+++ exporter/opentelemetry-exporter-zipkin/tests/test_zipkin_exporter.py
@@ -246,8 +246,8 @@ class TestZipkinSpanExporter(unittest.TestCase):
                     "key_bool": "False",
                     "key_string": "hello_world",
                     "key_float": "111.22",
-                    "otel.status_code": "2",
-                    "otel.status_description": "Example description",
+                    "otel.status_code": "ERROR",
+                    "error": "Example description",
                 },
                 "debug": True,
                 "parentId": format(parent_id, "x"),
@@ -272,10 +272,7 @@ class TestZipkinSpanExporter(unittest.TestCase):
                 "duration": durations[1] // 10 ** 3,
                 "localEndpoint": local_endpoint,
                 "kind": span_kind,
-                "tags": {
-                    "key_resource": "some_resource",
-                    "otel.status_code": "1",
-                },
+                "tags": {"key_resource": "some_resource"},
                 "annotations": None,
             },
             {
@@ -289,7 +286,6 @@ class TestZipkinSpanExporter(unittest.TestCase):
                 "tags": {
                     "key_string": "hello_world",
                     "key_resource": "some_resource",
-                    "otel.status_code": "1",
                 },
                 "annotations": None,
             },
@@ -304,7 +300,6 @@ class TestZipkinSpanExporter(unittest.TestCase):
                 "tags": {
                     "otel.instrumentation_library.name": "name",
                     "otel.instrumentation_library.version": "version",
-                    "otel.status_code": "1",
                 },
                 "annotations": None,
             },
@@ -383,7 +378,7 @@ class TestZipkinSpanExporter(unittest.TestCase):
                 "duration": duration // 10 ** 3,
                 "localEndpoint": local_endpoint,
                 "kind": SPAN_KIND_MAP_JSON[SpanKind.INTERNAL],
-                "tags": {"otel.status_code": "1"},
+                "tags": {},
                 "annotations": None,
                 "debug": True,
                 "parentId": "0aaaaaaaaaaaaaaa",
@@ -737,6 +732,7 @@ class TestZipkinSpanExporter(unittest.TestCase):
         otel_spans[1].resource = Resource(
             attributes={"key_resource": "some_resource"}
         )
+        otel_spans[1].set_status(Status(StatusCode.OK))
         otel_spans[1].end(end_time=end_times[1])
 
         otel_spans[2].start(start_time=start_times[2])
@@ -775,8 +771,8 @@ class TestZipkinSpanExporter(unittest.TestCase):
                         "key_bool": "False",
                         "key_string": "hello_world",
                         "key_float": "111.22",
-                        "otel.status_code": "2",
-                        "otel.status_description": "Example description",
+                        "otel.status_code": "ERROR",
+                        "error": "Example description",
                     },
                     debug=True,
                     parent_id=ZipkinSpanExporter.format_pbuf_span_id(
@@ -809,7 +805,7 @@ class TestZipkinSpanExporter(unittest.TestCase):
                     kind=span_kind,
                     tags={
                         "key_resource": "some_resource",
-                        "otel.status_code": "1",
+                        "otel.status_code": "OK",
                     },
                 ),
                 zipkin_pb2.Span(
@@ -825,7 +821,6 @@ class TestZipkinSpanExporter(unittest.TestCase):
                     tags={
                         "key_string": "hello_world",
                         "key_resource": "some_resource",
-                        "otel.status_code": "1",
                     },
                 ),
                 zipkin_pb2.Span(
@@ -841,7 +836,6 @@ class TestZipkinSpanExporter(unittest.TestCase):
                     tags={
                         "otel.instrumentation_library.name": "name",
                         "otel.instrumentation_library.version": "version",
-                        "otel.status_code": "1",
                     },
                 ),
             ],